Proyectos_y_Ejercicios/validador_datos/version_7/requirements.txt
pandas>=1.5.0
pyyaml>=6.0
